// src/types/api/ValSection.ts
export interface ValSection {
  groupId: number;
  sectionText: string | null;
  displayOrder: number | null;
  defaultColWidth1: number | null;
  defaultColWidth2: number | null;
  defaultColWidth3: number | null;
  defaultColWidth4: number | null;
  defaultColType1: string | null;
  defaultColType2: string | null;
  defaultColType3: string | null;
  defaultColType4: string | null;
  autoIndent: boolean | null;
}